Refactor projection.rs into module directory
Split the 2821-line projection.rs into projection/{mod,cartesian,polar}.rs
so map projection rendering can live in its own file.
